Changed Consistency to Consistency_check and Consistency_compare, to take both cases into account
