+++ etc/consistency_check.yaml
@@ -1,0 +1,22 @@
+database:
+  engine: "postgres://tester:password@localhost:5432/test"
+  tablenames:
+    file:
+      schema: null
+      table: loc_inst_files
+    event:
+      schema: null
+      table: loc_inst_gen2_file_events
+  echo: false
+  pool_class: QueuePool
+consistency:
+  plugin:
+    name: Gen3Butler
+    config:
+      root: /data/gen3repo
+      config: null
+      config_file: null
+logging:
+  file: null
+  format: "%(asctime)s:%(name)s:%(levelname)s:%(message)s"
+  level: INFO
