+++ etc/consistency_compare.yaml
@@ -1,0 +1,20 @@
+database:
+  engine: "postgres://tester:password@localhost:5432/test"
+  tablenames:
+    file:
+      schema: null
+      table: loc_inst_files
+    event:
+      schema: null
+      table: loc_inst_gen2_file_events
+  echo: false
+  pool_class: QueuePool
+consistency:
+  storage_source_one: database
+  source_one_data: "postgres://tester:password@localhost:5432/test"
+  storage_source_two: Gen3Butler
+  source_two_data: /data/gen3repo
+logging:
+  file: null
+  format: "%(asctime)s:%(name)s:%(levelname)s:%(message)s"
+  level: INFO
